Rewrite move encoding (resolves #103)
Rewrite move encoding of piece drops and king promotions
in order to fit into 16 bits, so that TT works properly.

The bits for the move type are set as if it was a normal move,
and the bits for the promotion piece type are used to set
flags for piece drops and king promotions.

The dropped piece is encoded in the bits for the origin square.