Improve logic for crazyhouse quad
Use only three arrays to tune for zh:
- quadOurs
- quadTheirs
- quadOursInHand

The idea is to keep the intent of us/them tables as before
which is for synergy of play and for domination of certain
piece types over opponent pieces, but to also add a synergy
for attacking ability based on certain pieces in hand. The
quadThemsInHand table doesn't really make sense.